Merge Blue Whale BMC support with latest OpenBMC upstream content
This set of code changes enables the Blue Whale BMC build with
the latest OpenBMC upstream content.  There was a set of files
that were simply copied from "master" branch over to this "mlnx-1.0"
branch, so those files are not part of this code review.  Please
review commit 7003c39 on the "mlnx-1.0" branch to see that file list.

The changes in this delivery are:
1) Modify recipes to use our bu-gerrit Git repos for Linux and U-Boot
2) Update Blue Whale's "bblayers.conf.sample" file to track changes
   in the layering scheme used in upstream OpenBMC
3) Update Blue Whale's "bluewhale-ast2500.conf" file to enable build
   of uImage kernel format via KERNEL_EXTRA_ARGS, and to declare
   the set of flash offsets for our image-bmc composite file
4) Changes to allow 'u-boot-env.bin' to be exported and then used
   in the packaging of the image-bmc composite file

NOTE: the following recipes are NOT part of the Blue Whale BMC image:
     a) MFT tools
     b) BMC tools
     c) IPMI daemon
     d) Fan daemon
Follow-on deliveries will fix the build errors found in these recipes.

NOTE2: while this image is not ready for public use, there are
some important changes to note with this delivery:
     GCC: changed from 5.3.0 to 7.3.0
     GNU binutils: changed from 2.26.0.20160214 to 2.30.0.20180208
     /etc/os-release: changed from 6.0.0200-XXX to v2.3-XXX
                      (we can certainly modify this per our needs)

Issue:1522059
Change-Id: Ia0b7e171464b128e8c85c884ede07b21484e8b7f