Change BMC Jenkins script to run U-Boot & Linux verify jobs remotely
When the kernel was switched to use "dev-4.17" branch, the tool
"bison" is now used during the DTB creation.  This tool is not
present on the jenkins server, therefore the bmc-linux verify
job fails when applied to the "dev-4.17" branch.  The change
here is to run both U-Boot and Linux verify jobs on the remote
host (mtbu-yocto in our case) to have access to all necessary
tools and to run on the same server as the OpenBMC nightly and
verify jobs.

Issue: 1526989
Change-Id: I6457d49b2df595cb66fb8fc50999d7fd9cc33132