Remove the use of "u-boot-env.bin" in creating "image-bmc"
This code delivery changes the way "image-bmc" is built so
that it does not depend on the "u-boot-env.bin" file.
This aligns the build of Blue Whale BMC with upstream,
and allows us to revert the file
  meta-phosphor/classes/image_types_phosphor.bbclass
so that it is now the same as upstream.

Also, revert the changes made to "linux-aspeed_git.bb",
this will help the next merge we do with upstream.

Issue: 1592398
Change-Id: Ic60d91d3fbcbc0a4f56a2f91baf3d9f6a1b56f9e