Modify 'update_all' script to handle downgrade to BMC with uImage format
During 'update_all' execution, the U-Boot env vars are modified and the
assumption was that the new image needed a boot command like:
  bootm 0x20070000
because the image being programmed was in fitImage format. However, if
there is a downgrade to a BMC with uImage format, this image actually
needs the bootm command with 3 args: 0x20070000 0x20360000 0x20350000

To fix this, I added logic to check the image type (fitImage or uImage)
of the image to be programmed, and thereafter set 'bootcmd_string'
accordingly.

Issue:1627236
Change-Id: Ia65c1746387e52c733f3a6a251433eb61ee7f252