Add ipmitool support to OpenBMC
ipmitool is needed on OpenBMC to be able to test the BMC and
BlueField IPMB drivers.
This change goes along with the bmc linux change associated
with RM ttask number: 1551766

We do 2 things here:
1) Support ipmitool on OpenBMC
2) Add an ipmitool patch to enable the correct functioning
of the BMC IPMB driver. This patch is needed because there
is no support on ipmitool to send an "IPMB" message from
the BMC without bridging.

Change-Id: I23539734382b4676a08e4c2aae1d85c728999a42