Bump up date in the Blue Whale BMC README file
We track the BMC image version as a whole based on the OpenBMC version
(e.g. "v2.4-XXX").  The "XXX" number in the OpenBMC version will only
increment if there is a code delivery to the OpenBMC git repo.
For example, fixing a bug with only a U-Boot delivery will create a new
image with the fix BUT it will have the same OpenBMC version as the
prior build.  This delivery is a short term fix to this problem.
I'm open to suggestions as to how to fix this issue long term.

Change-Id: I579c283ef284b1a50031832f745636b841d743fe