Fix the BMC IPMB ci
The BMC IPMB ci test fails because the ipmb_host module
fails to load at boot time. It fails to load because the
BF ipmb daemon is not running; it has been disabled at
boot time, and is enabled only if done so by the user.

To fix this, we need to disable loading the ipmb_host
driver on the BMC and only load it when running the ci
test.

Change-Id: If642c0d34804df462c9f91ac02fcd5f4e6704914