Support ipmb_dev_int driver
A customer has requested for a new feature where the BF is the
IPMB requester and the BMC is the responder. So the BMC needs
to be able to load the ipmb_dev_int driver to receive requests
and send the responses back.

Just like for the BF, I have created 2 scripts:
   1) The default mode at boot time, is the existing one where
      the BMC sends IPMB requests to the BF. No need to do
      any additional steps for that.
   2) if the customer wants to switch to the mode where he can
      send requests from the BF to the BMC, he can use the
      load_bf2bmc_ipmb.sh script. This script loads the
      ipmb_dev_int driver which was already implemented and
      upstreamed for the BF.
   3) if the customer wants to switch back to the default mode
      where the BMC sends requests to the BF, he can use
      the load_bmc2bf_ipmb.sh script.

I have also updated OpenIPMI so that it can handle IPMB
messages. This is the same code that I have added to the BF
and have upstreamed.

Change-Id: Id47f6a0be80215bda30d1e84bbc467483de1dd47