OpenIPMI: set sensors' name's size <= 16
In the latest ipmitool version, the size of the
sensors' names is limited to 16 characters.
We observe this when issuing "ipmitool sdr list"
or "ipmitool sensor list" commands from the BF to
the BMC. The last sensor: "vr_bfddr_1p2v_iout" is
18 characters long, so the ipmitool program only
prints "vr_bfddr_1p2v_io". I have changed the name
to "bfddr_1p2v_iout".

Change-Id: I4313690c33cf6e73c14283641195c3f3c1c67851