Add 'screen' program to the BMC image
The 'screen' program was present in the older BMC images,
i.e. version 6.0.0200-XXX, but is not present in the
latest BMC image.

NOTE: there are two ways to use the 'screen' program on
the BMC to attach to the BlueField's serial console:
   1) screen /dev/ttyS0 115200
   2) screen /dev/rshim0/console 115200

The fix here is to add the 'screen' recipe to our BMC's
OBMC_IMAGE_EXTRA_INSTALL_append definition.

Change-Id: I88d59794a61e312a9878088cce198a8c880c3a94