Load both ipmb host and client drivers at boot time
This change is dependent on its counter part in linux as
well as on the patch to support dual mode on the BF.
Now that I2C dual slave mode is supported on the AST2500,
both the ipmb_dev_int driver and the ipmb_host driver
can be loaded by default at boot time.
The mlx_ipmid.service is dependent on the ipmb_dev_int
driver. If ipmb_dev_int driver fails to load, the
mlx_ipmid will fail as well.

Change-Id: I2ae02cde1323e34c43cb7b617d33d0529d2b8740