Fix IPMI symlinks for sensor and fru collection
Issuing "ipmitool sdr list" command issued from either the BlueField
or an external host to the BMC shows sensors with 0 values.
This happens because the symlinks that OpenIPMI uses to collect sensors'
data are not being created. These symlinks are created by udev rules in:
70-mlxbw-hwmon-events.rules

This is probably due to some timing issue since if the BMC is rebooted
several times, some (not all, sometimes all) symlinks are created.

The above rule file does a match against SUBSYSTEM==hwmon and
DEVPATH==*hwmon*. The match probably does not happen so the action is not
executed. The following changes fixed the issue:
- merged 70-mlxbw-hwmon-events.rules and 70-mlxbw-events.rules into one
  file. It makes things simpler as well.
- moved 70-mlxbw-events.rules to 10-mlxbw-events.rules since the udev rules
  are executed in lexical order, i.e. higher priority for higher numbers.
  Now, this rule has the lowest priority.
- match SUBSYSTEM==i2c instead of hwmon because hwmon is anyways under
  i2c in this case. The match works with i2c as the subsystem.
- deleted DEVPATH and move the path directly as an argument to
  /usr/bin/mlxbw-hw-events.sh

This issue was observed in the past but was exacerbated after
adding the 99-fuse.rule (change brought by user space rshim)

Change-Id: I7dabb7fd11e0c59267682928653829b7c4c1fcff